final_project/requirements.txt
mypy
pytest
fpdf2
pick
click
tabulate